+++ service/controller/control_plane.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha2"
+	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
 	"github.com/giantswarm/apiextensions/v3/pkg/clientset/versioned"
 	"github.com/giantswarm/certs/v3/pkg/certs"
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
@@ -109,7 +109,7 @@ func NewControlPlane(config ControlPlaneConfig) (*ControlPlane, error) {
 			K8sClient: config.K8sClient,
 			Logger:    config.Logger,
 			NewRuntimeObjectFunc: func() runtime.Object {
-				return new(infrastructurev1alpha2.AWSControlPlane)
+				return new(infrastructurev1alpha3.AWSControlPlane)
 			},
 			Resources: resources,
 
@@ -517,16 +517,16 @@ func newControlPlaneResources(config ControlPlaneConfig) ([]resource.Interface, 
 	return resources, nil
 }
 
-func newControlPlaneToClusterFunc(g8sClient versioned.Interface) func(ctx context.Context, obj interface{}) (infrastructurev1alpha2.AWSCluster, error) {
-	return func(ctx context.Context, obj interface{}) (infrastructurev1alpha2.AWSCluster, error) {
+func newControlPlaneToClusterFunc(g8sClient versioned.Interface) func(ctx context.Context, obj interface{}) (infrastructurev1alpha3.AWSCluster, error) {
+	return func(ctx context.Context, obj interface{}) (infrastructurev1alpha3.AWSCluster, error) {
 		cr, err := key.ToControlPlane(obj)
 		if err != nil {
-			return infrastructurev1alpha2.AWSCluster{}, microerror.Mask(err)
+			return infrastructurev1alpha3.AWSCluster{}, microerror.Mask(err)
 		}
 
-		m, err := g8sClient.InfrastructureV1alpha2().AWSClusters(cr.Namespace).Get(ctx, key.ClusterID(&cr), metav1.GetOptions{})
+		m, err := g8sClient.InfrastructureV1alpha3().AWSClusters(cr.Namespace).Get(ctx, key.ClusterID(&cr), metav1.GetOptions{})
 		if err != nil {
-			return infrastructurev1alpha2.AWSCluster{}, microerror.Mask(err)
+			return infrastructurev1alpha3.AWSCluster{}, microerror.Mask(err)
 		}
 
 		return *m, nil