+++ service/controller/resource/tccpvpcidstatus/create.go
@@ -3,7 +3,7 @@ package tccpvpcidstatus
 import (
 	"context"
 
-	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha2"
+	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/operatorkit/v4/pkg/controller/context/reconciliationcanceledcontext"
 	"k8s.io/apimachinery/pkg/types"
@@ -18,7 +18,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		return microerror.Mask(err)
 	}
 
-	var cr infrastructurev1alpha2.AWSCluster
+	var cr infrastructurev1alpha3.AWSCluster
 	{
 		cl, err := key.ToCluster(ctx, obj)
 		if err != nil {