+++ service/internal/cloudconfig/tccpn_extension.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"encoding/base64"
 
-	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha2"
+	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
 	"github.com/giantswarm/certs/v3/pkg/certs"
 	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v10/pkg/template"
 	"github.com/giantswarm/microerror"
@@ -25,7 +25,7 @@ type TCCPNExtension struct {
 	awsCNIWarmIPTarget    string
 	baseDomain            string
 	cc                    *controllercontext.Context
-	cluster               infrastructurev1alpha2.AWSCluster
+	cluster               infrastructurev1alpha3.AWSCluster
 	clusterCerts          []certs.File
 	encrypter             encrypter.Interface
 	encryptionKey         string