+++ service/internal/images/images_test.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 	"testing"
 
-	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha2"
+	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
 	releasev1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/release/v1alpha1"
 	"github.com/giantswarm/operatorkit/v4/pkg/controller/context/cachekeycontext"
 
@@ -50,7 +50,7 @@ func Test_Images_Cache(t *testing.T) {
 				}
 			}
 
-			var cl infrastructurev1alpha2.AWSCluster
+			var cl infrastructurev1alpha3.AWSCluster
 			{
 				cl = unittest.DefaultCluster()
 			}