+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Introducing `v1alpha3` CR's.
+
 ## [10.6.1] - 2021-07-01
 
 ## Changed